fix validation, middleware add
