+++ middlewares/index.js
@@ -1,0 +1,3 @@
+import validateBody from "./validateBody";
+
+module.exports = { validateBody };
