+++ middlewares/validateBody.js
@@ -1,0 +1,15 @@
+const { HttpError } = require("../helpers");
+
+const validateBody = (schema) => {
+  const func = (req, res, next) => {
+    const { error } = schema.validate(req.body);
+    if (error) {
+      next(HttpError(400, error.message));
+    }
+    next();
+  };
+
+  return func;
+};
+
+module.exports = validateBody;
