+++ models/contacts.json
@@ -63,6 +63,12 @@
     "id": "ZgGyMR2KyoPHflFdAIO5D",
     "name": "Yan Kondratenko",
     "email": "kondyan.asusnexus@gmail.com",
-    "phone": "+38 (050) 624-40-94"
+    "phone": "+38 (050) 524-40-94"
+  },
+  {
+    "id": "yXYM9EBHOPuDPYgAQuNo5",
+    "name": "Yan",
+    "email": "kondyan@gmail.com",
+    "phone": "+38 (050) 24-40-94"
   }
 ]
